+++ src/actions/index.ts
@@ -8,7 +8,7 @@ import "./google/google_cloud_storage"
 import "./hipchat/hipchat"
 import "./jira/jira"
 import "./marketo/marketo"
-import "./queueaction/queueaction"
+import "./queueaction/queue_test_action"
 import "./segment/segment"
 import "./segment/segment_group"
 import "./segment/segment_track"
