+++ src/actions/segment/test_segment_group.ts
@@ -5,7 +5,7 @@ import * as Hub from "../../hub"
 import { SegmentGroupAction } from "./segment_group"
 
 const action = new SegmentGroupAction()
-action.runInOwnProcess = false
+action.executeInOwnProcess = false
 
 function expectSegmentMatch(request: Hub.ActionRequest, match: any) {
   const groupSpy = sinon.spy()
